------------------------------------------------------------------------ r1122 | bonro01 | 2009-01-28 14:46:56 -0500 (Wed, 28 Jan 2009) | 11 lines
INTEGRATION: Change 492790 in ingres on 22-jul-2008 by boija02 on main_boija02_gx270

Add build and run scripts for VMS and supporting files to allow
building against CLI and unixodbc. Fixed SEPPARAM_NODE usage, which is
defined only as a hostname, rather than a full vnode, replacing it with
SEPPARAM_ODB. Added a wrapper to the include of malloc.h, which isn't
present on VMS. Added a VMS-compatible cfg file to allow the suite to
be run using executor. SEPPARAM_ODB change requires changes to runodbc.sh
and runodbc.bat to match.

------------------------------------------------------------------------