------------------------------------------------------------------------ r1125 | bonro01 | 2009-01-28 14:47:45 -0500 (Wed, 28 Jan 2009) | 21 lines
INTEGRATION: Change 493981 in ingres on 10-oct-2008 by vande02 on main_usilqlnx9

Cross-integrating ingres2006r3 Change 493980 to MAIN.

Test cleanup for UTF8 in Replicator Part 3.

The main change to each test was to add conditional canons to allow the
test to run clean on UTF8 when the order of replicator catalogs or other
rows comes out in the UTF8 Unicode collation sequence order.

<< IF (ING_CHARSET == "UTF8")
...this is the UTF8 results
<<
...regular results
>>

The above is making the assumption the tester has noticed the variable in
the ingbenv.sh/ingenv.bat template named ING_CHARSET and has set it to the
appropriate charset under test = UTF8.

------------------------------------------------------------------------
r1126 | bonro01 | 2009-01-28 14:47:56 -0500 (Wed, 28 Jan 2009) | 17 lines

INTEGRATION: Change 494060 in ingres on 16-oct-2008 by hanal04 on hanal0493_intlnx_m1

Cross integration of change 494046

Added cuupdabfsrc field to the CURECORD and initialise it to FALSE
for all record types. This field must only be set to TRUE if the
first token in a record represents an abf_source value that should match
the object_id of the object.

When reloading objects during a "copyapp in" set cuupdabfsrc to TRUE
for the first detail record of an OC_AFORMREF object so that
IICUltdLineToDescriptor() can update the abf_source reference in-line
with the object's new object_id.

Update canons in abf61.sep to match the updated abf_source references.

------------------------------------------------------------------------