------------------------------------------------------------------------ r1126 | bonro01 | 2009-01-28 14:47:56 -0500 (Wed, 28 Jan 2009) | 17 lines
INTEGRATION: Change 494060 in ingres on 16-oct-2008 by hanal04 on hanal0493_intlnx_m1

Cross integration of change 494046

Added cuupdabfsrc field to the CURECORD and initialise it to FALSE
for all record types. This field must only be set to TRUE if the
first token in a record represents an abf_source value that should match
the object_id of the object.

When reloading objects during a "copyapp in" set cuupdabfsrc to TRUE
for the first detail record of an OC_AFORMREF object so that
IICUltdLineToDescriptor() can update the abf_source reference in-line
with the object's new object_id.

Update canons in abf61.sep to match the updated abf_source references.

------------------------------------------------------------------------
r1127 | bonro01 | 2009-01-28 14:48:08 -0500 (Wed, 28 Jan 2009) | 5 lines

INTEGRATION: Change 494102 in ingres on 16-oct-2008 by sarjo01 on main_sarjo01lnx

Change C variable to varbyte to make UTF8 compatible

------------------------------------------------------------------------