------------------------------------------------------------------------ r1137 | bonro01 | 2009-01-28 14:51:39 -0500 (Wed, 28 Jan 2009) | 9 lines
INTEGRATION: Change 494616 in ingres on 05-nov-2008 by vande02 on main_usilqlnx9

Test maintenance for MAIN only at this time.

qp304, 306, 309, utl48.log - Nightly build and run of HOQA showed up some
new logs due to MAIN having change to sequence objects which now support
bigint (SIR 120452).

------------------------------------------------------------------------
r1138 | bonro01 | 2009-01-28 14:51:58 -0500 (Wed, 28 Jan 2009) | 45 lines

INTEGRATION: Change 494621 in ingres on 05-nov-2008 by vande02 on main_usilqlnx9

Cross-integrating ingres2006r3 Change 494620 to MAIN.

Test clean up for UTF8.

dt37 and dt38.sep - Removed one query which was locating the '31991'th
character which is invalid for UTF8.  Modified some of the queries slightly
so the results would be consistent for UTF8 and non-UTF8.

dt39.sep - Changed one query which was locating the '31999'th character
which is invalid for UTF8 to the 15000th char.  Modified some of the queries
slightly so the results would be consistent for UTF8 and non-UTF8.

dt40.sep - The test program was changed to accept an additional argument
SEPPARAM_CHARSET and set the max column size to 32010 for both UTF8 and
non-UTF8.  Some string functions must be coerced to another data type such
as varchar to allow them to return a value from a long varchar column,
therefore, the column length will return 16000 in UTF8 when it is really
32010.  Changed query locating the character at locations 31998, 31999,
32000, 32001, 32002 to the 15998,15998,16000. These locations would be
invalid for UTF8 after coercion to varchar.

All 5 tests -
1)  The test program was changed to accept an additional argument
SEPPARAM_CHARSET and set the max column size to 16K for UTF8 or 32K
(default) for non-UTF8.

2)  Added a conditional canon to allow the test to run clean on UTF8
when the answer is appropriately 16000 (max character column length for
UTF8, except long varchar which is 2GB, except when the string function
is coerced to varchar, then it is back to 16K).

<< IF (ING_CHARSET == "UTF8")
...this is the UTF8 results
>>
<<
...regular results
>>

The above is making the assumption the tester has noticed the variable in
the ingbenv.sh/ingenv.bat template named ING_CHARSET and has set it to the
appropriate charset under test = UTF8.

------------------------------------------------------------------------