------------------------------------------------------------------------ r1141 | bonro01 | 2009-01-28 14:52:33 -0500 (Wed, 28 Jan 2009) | 30 lines
INTEGRATION: Change 495109 in ingres on 08-dec-2008 by hanal04 on hanal0493_intlnx_m1

Cross integration of change 495092

Added a -no_rep option for COPYDB and UNLOADDB to allow
the user to exclude all replicator objects from the generated
copy.in.

Exclusions are identified as follows:

Tables:
The content of dd_support_tables is used to exclude replicator
tables.

Indexes:
The exclusion of indexes uses the excluded table IDs to
exclude the associated indexes.

Events:
The content of dd_events is used to exclude "most" replicator events.
However, not all replicator events are added to dd_events so we also
exlcude all events named LIKE 'dd_%server%'. The

Procedures:
The table names for the shadow table entries found in dd_support_tables
(E.g. tab100001sha) is used to generate a NOT IN list of procedure
names to exclude (E.g. tab100001shrm_).

------------------------------------------------------------------------
r1142 | bonro01 | 2009-01-28 14:52:43 -0500 (Wed, 28 Jan 2009) | 5 lines

INTEGRATION: Change 495259 in ingres on 16-dec-2008 by sarjo01 on main_sarjo01lnx

Removed one of the "destructive" queries, which is now allowed in 9.3

------------------------------------------------------------------------