------------------------------------------------------------------------ r1270 | vande02 | 2009-03-02 17:30:24 -0500 (Mon, 02 Mar 2009) | 13 lines
Cross-integrating MAIN Change 496212 to OPENSRCQA.

As per discussions with QA / Sustaining Engineering, the idea of not branching
ING_TSTs is going to be postponed.

As discussed, Filenames are being reverted to the original naming convention.

Regression tests are still to be run from the main codeline for all versions.
_regr filenames are used for tests which are not in HOQA, but are run directly
by the regression suite generator.  These are generally regression tests
for older releases.

------------------------------------------------------------------------