------------------------------------------------------------------------ r1461 | vande02 | 2009-04-21 20:46:15 -0400 (Tue, 21 Apr 2009) | 9 lines
Cross-integrating MAIN change 497116 to OPENSRCQA.

Wrapped VMS-specific items in conditional and changed order of VMS
build flags for all tests, corrected SEPPARAM_NODE/ODB confusion in
dbnet.bat and made some necessary changes to the VMS runodbc script, to
use uppercase parameters and reverse ordering of CLI and unixodbc
testing.

------------------------------------------------------------------------
r1462 | vande02 | 2009-04-21 22:01:03 -0400 (Tue, 21 Apr 2009) | 12 lines

Cross-integrating MAIN change 490902 to OPENSRCQA.

RUNBE.BAT - Added the C2 Security Auditing (ON and OFF) iisetres to
this Windows script for c2secure.  This matches how it is done for
runbe.sh on UNIX.  Also, did a tedious cleanup of this script by
removing all the end of line characters.  Editing the file in vi on
Windows made it look very messy and hard to read.  Some how, with
the last submission on this file, those characters remained in the file.
Tested this script on Windows after removing those end of line character
and it worked just fine.  Removed the welcome echo messages.

------------------------------------------------------------------------