------------------------------------------------------------------------ r1828 | vande02 | 2009-08-05 20:45:12 -0400 (Wed, 05 Aug 2009) | 12 lines
Cross-integrating MAIN change 499287 to OPENSRCQA.

Test maintenance Windows script DBCREATE.BAT.  This script was assuming
a variable for the VNODE was being set - %lback%, however, such a variable
was not being set any where.  The actual VNODE name for the lbnet loop back
tests is lback which is created during tstsetup.bat.  The presence of variable
%lback% could not be found in any script or setup template.  I remove the
% signs making lback a variable and just put the name of the VNODE in the
DBCREATE.BAT script which is called by RUNLBNET.BAT so DBCREATE.BAT works
properly.

------------------------------------------------------------------------