------------------------------------------------------------------------ r1923 | vande02 | 2009-09-03 10:45:13 -0400 (Thu, 03 Sep 2009) | 11 lines
Cross-integrating MAIN Change 500006 to opensrcqa.

Test script maintenance to the QA load script templates.  The setting
of variable SEP_COBOL_LD has had the potential to interfere with
initlar.sep for a long time.  This initialization SEP test for LAR
also creates loc1 to loc5 and extends lardb to those locations.
The Area path for these locations is messed up when SEP_COBOL_LD
happens to also be set incorrectly that is with single quotes and
no dollar sign in front of II_SYSTEM.

------------------------------------------------------------------------