------------------------------------------------------------------------ r2482 | vande02 | 2010-01-22 18:45:18 -0500 (Fri, 22 Jan 2010) | 8 lines
Cross-integrating MAIN change 502130 to OPENSRCQA.

Changed default value of IIDCB to 20 in the Unix and Windows templates.
For all embed tests, single row contained two-digit date 14-nov-18,
which was expected to resolve to 1918, but with IIDCB of 20 was
resolving to 2018, causing diffs.

------------------------------------------------------------------------
r2483 | vande02 | 2010-01-22 18:45:31 -0500 (Fri, 22 Jan 2010) | 5 lines

Cross-integrating MAIN Change 501867 to OPENSRCQA.

Disabled length checking for long types

------------------------------------------------------------------------