------------------------------------------------------------------------ r2674 | vande02 | 2010-03-17 13:30:19 -0400 (Wed, 17 Mar 2010) | 7 lines
Cross-integrating MAIN Change 502400 to OPENSRCQA.

Cardinality related sep changes for scalar sub-selects.

qp032 is in effect reverted to the prior version.

------------------------------------------------------------------------
r2675 | vande02 | 2010-03-17 13:30:27 -0400 (Wed, 17 Mar 2010) | 13 lines

Cross-integrating MAIN 502851 to OPENSRCQA.

This change caused test utl25 to DIFF because this test used multiple lines
of the same -r flag to test an optimizedb -zf file where the file had hundreds
of flags. I've amended this test to instead use a table with hundreds of
columns and have many -a flags.

Cross integration of change 502801

Change 502769 removed the input file opt1000.txt but not the command that
used it.

------------------------------------------------------------------------